+++ tinker_cookbook/tokendb_studio/ui/src/components/Badge.tsx
