+++ tinker_cookbook/tokendb_studio/ui/src/components/ContextBar.tsx
