+++ tinker_cookbook/tokendb_studio/ui/src/components/DataTable.tsx
