+++ tinker_cookbook/tokendb_studio/ui/src/components/StatCard.tsx
