+++ tinker_cookbook/tokendb_studio/ui/src/components/VisualFrame.tsx
