+++ tinker_cookbook/tokendb_studio/ui/src/screens/Chat.tsx
