+++ tinker_cookbook/tokendb_studio/ui/src/screens/Dashboard.tsx
