+++ tinker_cookbook/tokendb_studio/ui/src/screens/Visuals.tsx
